Check Point: Compiles to about 75%

+++ engine/server/library/serverPathfinding/src/shared/CityPathGraph.cpp
@@ -709,7 +709,7 @@ void CityPathGraph::setLinkDistance ( float dist )
 // we cast the addresses of all the neighbors to ints, multiply them
 // by a random large prime number, then xor the bits together.
 
-int CityPathGraph::getNeighborCode ( int whichNode ) const
+int64_t CityPathGraph::getNeighborCode ( int whichNode ) const
 {
 	CityPathNode const * node = _getNode(whichNode);
 
@@ -717,7 +717,7 @@ int CityPathGraph::getNeighborCode ( int whichNode ) const
 
 	int edgeCount = node->getEdgeCount();
 
-	int code = 0;
+	int32_t code = 0;
 
 	for(int i = 0; i < edgeCount; i++)
 	{
@@ -727,9 +727,9 @@ int CityPathGraph::getNeighborCode ( int whichNode ) const
 
 		if(neighbor == nullptr) continue;
 
-		int neighborInt = reinterpret_cast<int>(neighbor);
+		int64_t neighborInt = reinterpret_cast<int64_t>(neighbor);
 
-		int mungedInt = neighborInt * 1295183;
+		int64_t mungedInt = neighborInt * 1295183;
 
 		code ^= mungedInt;
 	}
